change file module_b.py

+++ useful_package/module_b.py
@@ -1,2 +1,2 @@
 def hyperbola(x):
-    return float(3/x)
+    return 3/x
